+++ src/main/java/com/JS/ApiPlayground/model/User.java
@@ -1,0 +1,26 @@
+package com.JS.ApiPlayground.model;
+
+import lombok.*;
+import javax.persistence.*;
+
+@Entity
+@Table(name = "Users")
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+@ToString
+@EqualsAndHashCode
+
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "user_id")
+    private long id;
+    @Column(name = "user_name")
+    private String name;
+
+    public User(String name) {
+        this.name = name;
+    }
+}
