+++ src/main/java/com/JS/ApiPlayground/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.JS.ApiPlayground.repository;
+
+import com.JS.ApiPlayground.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {}
