--- src/main/java/com/JS/RESTPlayground/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package com.JS.RESTPlayground.repository;
-
-import com.JS.RESTPlayground.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {}
